+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## Currently Unreleased - TBD
 
-* Support register member functions with noexcept-specification. (Only after C++17)
+* Support registering member functions with noexcept-specification for C++17 or later.
 
 ## v1.2.1 - 2024.10.15